+++ source/Dialogs/SettingsDialog.Designer.vb
@@ -28,6 +28,9 @@ Partial Class SettingsDialog
         Label_defaultLineWidth = New Label()
         Label_loggingLevel = New Label()
         GroupBox_Chart = New GroupBox()
+        Button_ShowFontDialog = New Button()
+        ComboBox_Font = New ComboBox()
+        Label_DeafultFont = New Label()
         NumericUpDown_DefaultFontSize = New NumericUpDown()
         Label_DefaultFontSize = New Label()
         GroupBox_Logging = New GroupBox()
@@ -35,6 +38,7 @@ Partial Class SettingsDialog
         GroupBox_Interface = New GroupBox()
         Label_showOverview = New Label()
         CheckBox_showOverviewOnStartup = New CheckBox()
+        FontDialog1 = New FontDialog()
         CType(NumericUpDown_DefaultLineWidth, ComponentModel.ISupportInitialize).BeginInit()
         GroupBox_Chart.SuspendLayout()
         CType(NumericUpDown_DefaultFontSize, ComponentModel.ISupportInitialize).BeginInit()
@@ -45,30 +49,30 @@ Partial Class SettingsDialog
         ' Button_Close
         ' 
         Button_Close.Anchor = AnchorStyles.Bottom Or AnchorStyles.Right
-        Button_Close.Location = New Point(181, 320)
+        Button_Close.Location = New Point(242, 320)
         Button_Close.Name = "Button_Close"
         Button_Close.Size = New Size(80, 27)
-        Button_Close.TabIndex = 2
+        Button_Close.TabIndex = 3
         Button_Close.Text = "Close"
         ' 
         ' NumericUpDown_DefaultLineWidth
         ' 
         NumericUpDown_DefaultLineWidth.Anchor = AnchorStyles.Top Or AnchorStyles.Left Or AnchorStyles.Right
-        NumericUpDown_DefaultLineWidth.Location = New Point(154, 52)
+        NumericUpDown_DefaultLineWidth.Location = New Point(147, 75)
         NumericUpDown_DefaultLineWidth.Maximum = New Decimal(New Integer() {10, 0, 0, 0})
         NumericUpDown_DefaultLineWidth.Minimum = New Decimal(New Integer() {1, 0, 0, 0})
         NumericUpDown_DefaultLineWidth.Name = "NumericUpDown_DefaultLineWidth"
-        NumericUpDown_DefaultLineWidth.Size = New Size(87, 23)
-        NumericUpDown_DefaultLineWidth.TabIndex = 3
+        NumericUpDown_DefaultLineWidth.Size = New Size(155, 23)
+        NumericUpDown_DefaultLineWidth.TabIndex = 6
         NumericUpDown_DefaultLineWidth.Value = New Decimal(New Integer() {1, 0, 0, 0})
         ' 
         ' Label_defaultLineWidth
         ' 
         Label_defaultLineWidth.AutoSize = True
-        Label_defaultLineWidth.Location = New Point(6, 54)
+        Label_defaultLineWidth.Location = New Point(6, 77)
         Label_defaultLineWidth.Name = "Label_defaultLineWidth"
         Label_defaultLineWidth.Size = New Size(135, 15)
-        Label_defaultLineWidth.TabIndex = 2
+        Label_defaultLineWidth.TabIndex = 5
         Label_defaultLineWidth.Text = "Default series line width:"
         ' 
         ' Label_loggingLevel
@@ -83,35 +87,67 @@ Partial Class SettingsDialog
         ' GroupBox_Chart
         ' 
         GroupBox_Chart.Anchor = AnchorStyles.Top Or AnchorStyles.Left Or AnchorStyles.Right
+        GroupBox_Chart.Controls.Add(Button_ShowFontDialog)
+        GroupBox_Chart.Controls.Add(ComboBox_Font)
+        GroupBox_Chart.Controls.Add(Label_DeafultFont)
         GroupBox_Chart.Controls.Add(NumericUpDown_DefaultFontSize)
         GroupBox_Chart.Controls.Add(Label_DefaultFontSize)
         GroupBox_Chart.Controls.Add(Label_defaultLineWidth)
         GroupBox_Chart.Controls.Add(NumericUpDown_DefaultLineWidth)
         GroupBox_Chart.Location = New Point(14, 81)
         GroupBox_Chart.Name = "GroupBox_Chart"
-        GroupBox_Chart.Size = New Size(247, 90)
-        GroupBox_Chart.TabIndex = 0
+        GroupBox_Chart.Size = New Size(308, 110)
+        GroupBox_Chart.TabIndex = 1
         GroupBox_Chart.TabStop = False
         GroupBox_Chart.Text = "Chart"
+        ' 
+        ' Button_ShowFontDialog
+        ' 
+        Button_ShowFontDialog.Anchor = AnchorStyles.Top Or AnchorStyles.Right
+        Button_ShowFontDialog.Location = New Point(277, 16)
+        Button_ShowFontDialog.Name = "Button_ShowFontDialog"
+        Button_ShowFontDialog.Size = New Size(25, 23)
+        Button_ShowFontDialog.TabIndex = 2
+        Button_ShowFontDialog.Text = "..."
+        Button_ShowFontDialog.UseVisualStyleBackColor = True
+        ' 
+        ' ComboBox_Font
+        ' 
+        ComboBox_Font.Anchor = AnchorStyles.Top Or AnchorStyles.Left Or AnchorStyles.Right
+        ComboBox_Font.DropDownStyle = ComboBoxStyle.DropDownList
+        ComboBox_Font.FormattingEnabled = True
+        ComboBox_Font.Location = New Point(147, 16)
+        ComboBox_Font.Name = "ComboBox_Font"
+        ComboBox_Font.Size = New Size(124, 23)
+        ComboBox_Font.TabIndex = 1
+        ' 
+        ' Label_DeafultFont
+        ' 
+        Label_DeafultFont.AutoSize = True
+        Label_DeafultFont.Location = New Point(6, 19)
+        Label_DeafultFont.Name = "Label_DeafultFont"
+        Label_DeafultFont.Size = New Size(73, 15)
+        Label_DeafultFont.TabIndex = 0
+        Label_DeafultFont.Text = "Default font:"
         ' 
         ' NumericUpDown_DefaultFontSize
         ' 
         NumericUpDown_DefaultFontSize.Anchor = AnchorStyles.Top Or AnchorStyles.Left Or AnchorStyles.Right
-        NumericUpDown_DefaultFontSize.Location = New Point(154, 22)
+        NumericUpDown_DefaultFontSize.Location = New Point(147, 45)
         NumericUpDown_DefaultFontSize.Maximum = New Decimal(New Integer() {20, 0, 0, 0})
         NumericUpDown_DefaultFontSize.Minimum = New Decimal(New Integer() {6, 0, 0, 0})
         NumericUpDown_DefaultFontSize.Name = "NumericUpDown_DefaultFontSize"
-        NumericUpDown_DefaultFontSize.Size = New Size(87, 23)
-        NumericUpDown_DefaultFontSize.TabIndex = 1
+        NumericUpDown_DefaultFontSize.Size = New Size(155, 23)
+        NumericUpDown_DefaultFontSize.TabIndex = 4
         NumericUpDown_DefaultFontSize.Value = New Decimal(New Integer() {10, 0, 0, 0})
         ' 
         ' Label_DefaultFontSize
         ' 
         Label_DefaultFontSize.AutoSize = True
-        Label_DefaultFontSize.Location = New Point(6, 24)
+        Label_DefaultFontSize.Location = New Point(6, 47)
         Label_DefaultFontSize.Name = "Label_DefaultFontSize"
         Label_DefaultFontSize.Size = New Size(95, 15)
-        Label_DefaultFontSize.TabIndex = 0
+        Label_DefaultFontSize.TabIndex = 3
         Label_DefaultFontSize.Text = "Default font size:"
         ' 
         ' GroupBox_Logging
@@ -119,10 +155,10 @@ Partial Class SettingsDialog
         GroupBox_Logging.Anchor = AnchorStyles.Top Or AnchorStyles.Left Or AnchorStyles.Right
         GroupBox_Logging.Controls.Add(Label_loggingLevel)
         GroupBox_Logging.Controls.Add(CheckBox_logShowDebugMessages)
-        GroupBox_Logging.Location = New Point(14, 178)
+        GroupBox_Logging.Location = New Point(14, 197)
         GroupBox_Logging.Name = "GroupBox_Logging"
-        GroupBox_Logging.Size = New Size(247, 60)
-        GroupBox_Logging.TabIndex = 1
+        GroupBox_Logging.Size = New Size(308, 60)
+        GroupBox_Logging.TabIndex = 2
         GroupBox_Logging.TabStop = False
         GroupBox_Logging.Text = "Logging"
         ' 
@@ -130,7 +166,7 @@ Partial Class SettingsDialog
         ' 
         CheckBox_logShowDebugMessages.Anchor = AnchorStyles.Top Or AnchorStyles.Right
         CheckBox_logShowDebugMessages.AutoSize = True
-        CheckBox_logShowDebugMessages.Location = New Point(226, 25)
+        CheckBox_logShowDebugMessages.Location = New Point(287, 25)
         CheckBox_logShowDebugMessages.Name = "CheckBox_logShowDebugMessages"
         CheckBox_logShowDebugMessages.Size = New Size(15, 14)
         CheckBox_logShowDebugMessages.TabIndex = 1
@@ -143,36 +179,42 @@ Partial Class SettingsDialog
         GroupBox_Interface.Controls.Add(CheckBox_showOverviewOnStartup)
         GroupBox_Interface.Location = New Point(14, 14)
         GroupBox_Interface.Name = "GroupBox_Interface"
-        GroupBox_Interface.Size = New Size(246, 60)
-        GroupBox_Interface.TabIndex = 3
+        GroupBox_Interface.Size = New Size(307, 60)
+        GroupBox_Interface.TabIndex = 0
         GroupBox_Interface.TabStop = False
         GroupBox_Interface.Text = "Interface"
         ' 
         ' Label_showOverview
         ' 
         Label_showOverview.AutoSize = True
-        Label_showOverview.Location = New Point(8, 23)
+        Label_showOverview.Location = New Point(6, 23)
         Label_showOverview.Name = "Label_showOverview"
         Label_showOverview.Size = New Size(176, 15)
-        Label_showOverview.TabIndex = 1
+        Label_showOverview.TabIndex = 0
         Label_showOverview.Text = "Show overview chart on startup:"
         ' 
         ' CheckBox_showOverviewOnStartup
         ' 
         CheckBox_showOverviewOnStartup.Anchor = AnchorStyles.Top Or AnchorStyles.Right
         CheckBox_showOverviewOnStartup.AutoSize = True
-        CheckBox_showOverviewOnStartup.Location = New Point(224, 23)
+        CheckBox_showOverviewOnStartup.Location = New Point(285, 23)
         CheckBox_showOverviewOnStartup.Name = "CheckBox_showOverviewOnStartup"
         CheckBox_showOverviewOnStartup.Size = New Size(15, 14)
-        CheckBox_showOverviewOnStartup.TabIndex = 0
+        CheckBox_showOverviewOnStartup.TabIndex = 1
         CheckBox_showOverviewOnStartup.UseVisualStyleBackColor = True
+        ' 
+        ' FontDialog1
+        ' 
+        FontDialog1.AllowVerticalFonts = False
+        FontDialog1.FixedPitchOnly = True
+        FontDialog1.ShowEffects = False
         ' 
         ' SettingsDialog
         ' 
         AcceptButton = Button_Close
         AutoScaleDimensions = New SizeF(7F, 15F)
         AutoScaleMode = AutoScaleMode.Font
-        ClientSize = New Size(273, 359)
+        ClientSize = New Size(334, 359)
         Controls.Add(GroupBox_Interface)
         Controls.Add(Button_Close)
         Controls.Add(GroupBox_Logging)
@@ -181,7 +223,7 @@ Partial Class SettingsDialog
         Margin = New Padding(4, 3, 4, 3)
         MaximizeBox = False
         MinimizeBox = False
-        MinimumSize = New Size(289, 398)
+        MinimumSize = New Size(350, 398)
         Name = "SettingsDialog"
         ShowInTaskbar = False
         StartPosition = FormStartPosition.CenterParent
@@ -210,4 +252,8 @@ Partial Class SettingsDialog
     Friend WithEvents GroupBox_Interface As GroupBox
     Friend WithEvents Label_showOverview As Label
     Friend WithEvents CheckBox_showOverviewOnStartup As CheckBox
+    Friend WithEvents ComboBox_Font As ComboBox
+    Friend WithEvents Label_DeafultFont As Label
+    Friend WithEvents FontDialog1 As FontDialog
+    Friend WithEvents Button_ShowFontDialog As Button
 End Class